+++ src/eventNames.ts
@@ -30,6 +30,11 @@ const SESSION_SYNTHETICS_EVENT_NAMES = [
   'channels:notify',
   'ended:fromserver',
   'main-cam-control',
+  'participant:added-to-list-moderators',
+  'participant:removed-from-list-moderators',
+  'participant:move-request-to-conference',
+  'participant:move-request-to-stream',
+  'participant:canceling-word-request',
 ] as const;
 
 export const SESSION_JSSIP_EVENT_NAMES = [